requirements.txt
openai>=1.0.0
edge-tts>=6.1.0
moviepy>=1.0.3
pillow>=10.0.0
python-dotenv>=1.0.0
